Fix routes that may match on params early and fail later
Fixes issue #2

[Bug]

Whenever we had a value from the path that would match a parameter, we
would immediately recurse into the next routing table. This is a
problem, because a routing table may contain multiple params, and we
don't really know if the table we just recursed into, will continue to
match later on, it mail not match when it encounters a regexp or base
path.

For example:

If we have this two routes:

`/:id/:second_id/users`
`/:id/:second_id/places`

The second route would never match, on a path such as:

`/123/123/places`

Since we would always match on the first route, and fail on the last
part of the path.

[Solution]

Save the last real table match in a temporary variable.

A "real table match" is any match that was performed on a "base" path,
since params may match any path, but may will later fail if any base
path does not match, thus we need to go back in the routing tree and try
the next param on the last table we had a "real" match.

[Notes]

I'm not satisfied with our code, we are maintaining way too much state,
we should probably write a more "functional" code, and write a recursive
function instead of retaining state in variables, but this will probably
affect our performance promises (although I don't really care anymore)....